LayoutTests/imported/w3c/web-platform-tests/svg/animations/repeatn-remove-add-animation-expected.txt


FAIL This removes and adds an animation element while the animation is repeating assert_approx_equals: 2.2 expected 0 +/- 2 but got 255
